jarvis-backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
automat==25.4.16
cbor2==5.8.0
cffi==2.0.0
channels==4.3.2
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
django==6.0.1
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
msgpack==1.1.2
packaging==25.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pyopenssl==25.3.0
service-identity==24.2.0
sqlparse==0.5.5
twisted==25.5.0
txaio==25.12.2
typing-extensions==4.15.0
ujson==5.11.0
zope-interface==8.2
